@@ -42,7 +42,7 @@ public class UpdateTripHeadsignByDestinationStrategy implements GtfsTransformStr
     for (Trip trip : dao.getAllTrips()) {
       List<StopTime> stopTimes = dao.getStopTimesForTrip(trip);
       if (stopTimes != null && stopTimes.size() > 0) {
-        String tripHeadSign = stopTimes.get(stopTimes.size() - 1).getStop().getName();
+        String tripHeadSign = stopTimes.getLast().getStop().getName();
         if (tripHeadSign != null) {
           trip.setTripHeadsign(tripHeadSign);
           update++;
