@@ -45,7 +45,7 @@ public class UpdateTripHeadsignByReference implements GtfsTransformStrategy {
     for (Trip trip : dao.getAllTrips()) {
       List<StopTime> stopTimes = dao.getStopTimesForTrip(trip);
       if (stopTimes != null && stopTimes.size() > 0) {
-        Stop stop = (Stop) stopTimes.get(stopTimes.size() - 1).getStop();
+        Stop stop = (Stop) stopTimes.getLast().getStop();
         Stop gtfsStop =
             reference.getStopForId(
                 new AgencyAndId(getReferenceAgencyId(reference), stop.getMtaStopId()));
